refactor: Simplify authentication handling by replacing custom AuthProvider with SessionProvider and enhance session logging
It wasn't working on production due to extra hardening that was in place, overriding the configuration and still allowing the user to configure secure cookies solves the issue.

Something I'm still not 100% is that next standalone doesn't seems to update the page component causing the client session to be out of sync with the server. Forcing update of the component once, helps to mitigate this issue. Will investigate it further, on development this is not an issue.

Signed-off-by: Eden Reich <eden.reich@gmail.com>