feat(storage): implement UUID generation and improve session handling in storage services
I'm letting the application generate the UUID instead of the database so we are more flexible in terms of adding more storages later and having uuid being consistent among different storages.

I'm also removing the hook, because the storage config is passed from the server and when using hooks the client is loaded sometimes with the default config and then with the server config causing inconsistency. Since we have a server component we can simply pass it to the client.

I still need to optimize the amount of calls done from the client to the database. On every query we get several of queries to the API that then ending up calling the database. Unfortunatly with postgres there is no client library that calls it directly.

Signed-off-by: Eden Reich <eden.reich@gmail.com>